First draft for restructuring GPU communications and geometry
