+++ TestProgram/Memory/check_sync_GPU.in
@@ -1,0 +1,25 @@
+// Global variables 
+GLB_T = 8 
+GLB_X = 4
+GLB_Y = 4
+GLB_Z = 4
+
+NP_T = 2
+NP_X = 1
+NP_Y = 1
+NP_Z = 1
+
+rlx_level = 1
+rlx_seed = 12345
+rlx_start = new
+rlx_state = rlx_statae
+
+// Replicas
+N_REP = 1
+
+//Logger levels (default = -1)
+log:default = -1
+log:inverter = -1
+log:forcestat = -1
+
+gpuID = 0
